+++ test/data.test.js
@@ -1,9 +1,9 @@
 import test from 'tape'
-import { find, convert } from '../src/index.js'
+import { learn, convert } from '../src/index.js'
 import future from './data/future-simple.js'
 
 test('future tense:', function (t) {
-  let model = find(future)
+  let model = learn(future)
   // test them all
   future.forEach((a) => {
     let created = convert(a[0], model)
@@ -13,7 +13,7 @@ test('future tense:', function (t) {
 })
 
 test('french-nous:', function (t) {
-  let model = find(future)
+  let model = learn(future)
   // test them all
   future.forEach((a) => {
     let created = convert(a[0], model)