+++ test/misc.test.js
@@ -1,5 +1,5 @@
 import test from 'tape'
-import { find } from '../src/index.js'
+import { learn, compress } from '../src/index.js'
 
 test('one rule:', function (t) {
   const words = [
@@ -7,7 +7,8 @@ test('one rule:', function (t) {
     ['smoke', 'smoking'],
     ['talk', 'talking'],
   ]
-  let res = find(words)
+  let res = learn(words)
+  // res = compress(res)
   t.equal(res.rules.k.length, 1, 'one rule')
   t.equal(res.rules.k[0][0], 'lk', 'lk')
   t.equal(res.rules.k[0][1], 'lking', 'lking')
@@ -22,7 +23,8 @@ test('two rules:', function (t) {
     ['create', 'creating'],
     ['talk', 'talking'],
   ]
-  let res = find(words)
+  let res = learn(words)
+  // res = compress(res)
   // t.equal(res.rules.k.length, 2, 'two rules')
   t.equal(res.rules.k[0][0], 'lk', 'lk')
   t.equal(res.rules.e[0][0], 'e', 'e')
@@ -30,13 +32,14 @@ test('two rules:', function (t) {
   t.end()
 })
 
-test('find append:', function (t) {
+test('learn append:', function (t) {
   const words = [
     ['walk', 'walking'],
     ['wait', 'waiting'],
     ['sing', 'singing'],
   ]
-  let res = find(words)
+  let res = learn(words)
+  // res = compress(res)
   t.equal(res.rules[''].length, 1, 'one rules')
   // t.equal(res.rules.k[0][0], '', 'empty from prefix')
   t.equal(res.rules[''][0][1], 'ing', 'ing')