+++ changelog.md
@@ -1,3 +1,9 @@
+
+### 3.0.0 [unreleased]
+- **[change]** - don't compress by default
+- **[change]** - rename 'find' to 'learn'
+- **[new]** - add uncompress method
+
 ### 2.0.0 [Aug 26 2021]
 
-- **[change]** - - add compression to model
+- **[change]** - add compression to model